test(EnhancedClient): Refactor InvalidAnnotatedNestedBean
