+++ DynamoDbEncryption/runtimes/java/src/test/java/software/amazon/cryptography/dbencryptionsdk/dynamodb/enhancedclient/InvalidAnnotatedNestedBean.java
@@ -34,8 +34,69 @@ public class InvalidAnnotatedNestedBean {
     public NestedBean getNestedBeanClass() {
         return this.nestedBeanClass;
     }
-
     public void setNestedBeanClass(NestedBean nestedBeanClass) {
         this.nestedBeanClass = nestedBeanClass;
+    }
+
+    @DynamoDbBean
+    public static class NestedBean {
+        private String id;
+        private String firstName;
+        private String lastName;
+
+        public NestedBean() {};
+
+        NestedBean(String id, String firstName, String lastName) {
+            this.id = id;
+            this.firstName = firstName;
+            this.lastName = lastName;
+        }
+
+        @DynamoDbAttribute("id")
+        public String getId() {
+            return this.id;
+        }
+
+        public void setId(String id) {
+            this.id = id;
+        }
+
+        @DynamoDbEncryptionSignOnly
+        @DynamoDbAttribute("firstName")
+        public String getFirstName() {
+            return firstName;
+        }
+
+        public void setFirstName(String firstName) {
+            this.firstName = firstName;
+        }
+
+        @DynamoDbEncryptionDoNothing
+        @DynamoDbAttribute("lastName")
+        public String getLastName() {
+            return lastName;
+        }
+
+        public void setLastName(String lastName) {
+            this.lastName = lastName;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            if (this == obj) return true;
+            if (obj == null) return false;
+            if (getClass() != obj.getClass()) return false;
+            NestedBean other = (NestedBean) obj;
+            if (id == null) {
+                if (other.getId() != null) return false;
+            } else if (!id.equals(other.getId())) return false;
+            if (firstName == null) {
+                if (other.getFirstName() != null) return false;
+            } else if (!firstName.equals(other.getFirstName())) return false;
+            if (lastName == null) {
+                if (other.getLastName() != null) return false;
+            } else if (!lastName.equals(other.getLastName())) return false;
+            return true;
+        }
     }
 }
